requirements.txt
numpy
pandas
scipy
scikit-learn
prompt-toolkit
pyserial
mne